Removed functions to read and write from IOEvent, implemented only set and isSignaled
Rename IOEvent to IOAsyncAutoResetEvent to improve code readability

+++ include/tinycoro/io/IOOperation.hpp
@@ -27,7 +27,7 @@ namespace tinycoro::io
 
         friend class IOContext;
 
-        int fd;
+        uint64_t fd;
         IOContext& ioContext;
         epoll_event settings;
     };
